+++ sdk/resourcemanager/computeschedule/armcomputeschedule/scheduledactionextension_client_example_test.go
@@ -11,8 +11,8 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2025-04-15-preview/ScheduledActionExtension_ListByVms_MaximumSet_Gen.json
-func ExampleScheduledActionExtensionClient_NewListByVMsPager_scheduledActionExtensionListByVmsMaximumSet() {
+// Generated from example definition: 2026-03-01-preview/ScheduledActionExtension_ListByVms_MaximumSet_Gen.json
+func ExampleScheduledActionExtensionClient_NewListByVMsPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -22,7 +22,7 @@ func ExampleScheduledActionExtensionClient_NewListByVMsPager_scheduledActionExte
 	if err != nil {
 		log.Fatalf("failed to create client: %v", err)
 	}
-	pager := clientFactory.NewScheduledActionExtensionClient().NewListByVMsPager("sazvpabfud", nil)
+	pager := clientFactory.NewScheduledActionExtensionClient().NewListByVMsPager("rgdhjh", nil)
 	for pager.More() {
 		page, err := pager.NextPage(ctx)
 		if err != nil {
@@ -40,11 +40,11 @@ func ExampleScheduledActionExtensionClient_NewListByVMsPager_scheduledActionExte
 		// 				Properties: &armcomputeschedule.ScheduledActionsExtensionProperties{
 		// 					ResourceType: to.Ptr(armcomputeschedule.ResourceTypeVirtualMachine),
 		// 					ActionType: to.Ptr(armcomputeschedule.ScheduledActionTypeStart),
-		// 					StartTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.281Z"); return t}()),
-		// 					EndTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.286Z"); return t}()),
+		// 					StartTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2026-03-12T02:39:41.921Z"); return t}()),
+		// 					EndTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2026-03-12T02:39:41.921Z"); return t}()),
 		// 					Schedule: &armcomputeschedule.ScheduledActionsSchedule{
-		// 						ScheduledTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.TimeOnly, "19:00:00"); return t}()),
-		// 						TimeZone: to.Ptr("g"),
+		// 						ScheduledTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.TimeOnly, "12:00:00"); return t}()),
+		// 						TimeZone: to.Ptr("ywgarwvwqegtdsdvv"),
 		// 						RequestedWeekDays: []*armcomputeschedule.WeekDay{
 		// 							to.Ptr(armcomputeschedule.WeekDayMonday),
 		// 						},
@@ -52,20 +52,21 @@ func ExampleScheduledActionExtensionClient_NewListByVMsPager_scheduledActionExte
 		// 							to.Ptr(armcomputeschedule.MonthJanuary),
 		// 						},
 		// 						RequestedDaysOfTheMonth: []*int32{
-		// 							to.Ptr[int32](15),
+		// 							to.Ptr[int32](18),
 		// 						},
 		// 						ExecutionParameters: &armcomputeschedule.ExecutionParameters{
 		// 							OptimizationPreference: to.Ptr(armcomputeschedule.OptimizationPreferenceCost),
 		// 							RetryPolicy: &armcomputeschedule.RetryPolicy{
-		// 								RetryCount: to.Ptr[int32](17),
-		// 								RetryWindowInMinutes: to.Ptr[int32](29),
+		// 								RetryCount: to.Ptr[int32](19),
+		// 								RetryWindowInMinutes: to.Ptr[int32](3),
+		// 								OnFailureAction: to.Ptr(armcomputeschedule.ResourceOperationTypeUnknown),
 		// 							},
 		// 						},
 		// 						DeadlineType: to.Ptr(armcomputeschedule.DeadlineTypeUnknown),
 		// 					},
 		// 					NotificationSettings: []*armcomputeschedule.NotificationProperties{
 		// 						{
-		// 							Destination: to.Ptr("wbhryycyolvnypjxzlawwvb"),
+		// 							Destination: to.Ptr("zaaoabozbhyuhejwsrennfsxowp"),
 		// 							Type: to.Ptr(armcomputeschedule.NotificationTypeEmail),
 		// 							Language: to.Ptr(armcomputeschedule.LanguageEnUs),
 		// 							Disabled: to.Ptr(true),
@@ -75,7 +76,7 @@ func ExampleScheduledActionExtensionClient_NewListByVMsPager_scheduledActionExte
 		// 					ProvisioningState: to.Ptr(armcomputeschedule.ProvisioningStateSucceeded),
 		// 					ResourceNotificationSettings: []*armcomputeschedule.NotificationProperties{
 		// 						{
-		// 							Destination: to.Ptr("hjnstepnzzdg"),
+		// 							Destination: to.Ptr("zaaoabozbhyuhejwsrennfsxowp"),
 		// 							Type: to.Ptr(armcomputeschedule.NotificationTypeEmail),
 		// 							Language: to.Ptr(armcomputeschedule.LanguageEnUs),
 		// 							Disabled: to.Ptr(true),
@@ -83,97 +84,15 @@ func ExampleScheduledActionExtensionClient_NewListByVMsPager_scheduledActionExte
 		// 					},
 		// 				},
 		// 				ID: to.Ptr("/subscriptions/83C27AB3-A7B9-498B-B165-D9440661474F/resourceGroups/myRg/providers/Microsoft.ComputeSchedule/scheduledActions/myScheduledAction"),
-		// 				Name: to.Ptr("lktvylsmyrhgllrxisfdeqd"),
-		// 				Type: to.Ptr("afmfvpvaoczrktku"),
+		// 				Name: to.Ptr("myScheduledAction"),
+		// 				Type: to.Ptr("pbkwnfqervqhogxhemmgknol"),
 		// 				SystemData: &armcomputeschedule.SystemData{
-		// 					CreatedBy: to.Ptr("cvryvreuvvjtiamcwhisrt"),
+		// 					CreatedBy: to.Ptr("oadcwbhkptyfimpbaoyf"),
 		// 					CreatedByType: to.Ptr(armcomputeschedule.CreatedByTypeUser),
-		// 					CreatedAt: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.288Z"); return t}()),
-		// 					LastModifiedBy: to.Ptr("supbnksztdbgulxgvfmqvriqdlpirh"),
+		// 					CreatedAt: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2026-03-12T02:39:41.641Z"); return t}()),
+		// 					LastModifiedBy: to.Ptr("djuqozdplhpcdifwrsetsuu"),
 		// 					LastModifiedByType: to.Ptr(armcomputeschedule.CreatedByTypeUser),
-		// 					LastModifiedAt: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.288Z"); return t}()),
-		// 				},
-		// 			},
-		// 		},
-		// 		NextLink: to.Ptr("https://microsoft.com/a"),
-		// 	},
-		// }
-	}
-}
-
-// Generated from example definition: 2025-04-15-preview/ScheduledActionExtension_ListByVms_MinimumSet_Gen.json
-func ExampleScheduledActionExtensionClient_NewListByVMsPager_scheduledActionExtensionListByVmsMinimumSet() {
-	cred, err := azidentity.NewDefaultAzureCredential(nil)
-	if err != nil {
-		log.Fatalf("failed to obtain a credential: %v", err)
-	}
-	ctx := context.Background()
-	clientFactory, err := armcomputeschedule.NewClientFactory("<subscriptionID>", cred, nil)
-	if err != nil {
-		log.Fatalf("failed to create client: %v", err)
-	}
-	pager := clientFactory.NewScheduledActionExtensionClient().NewListByVMsPager("sazvpabfud", nil)
-	for pager.More() {
-		page, err := pager.NextPage(ctx)
-		if err != nil {
-			log.Fatalf("failed to advance page: %v", err)
-		}
-		for _, v := range page.Value {
-			// You could use page here. We use blank identifier for just demo purposes.
-			_ = v
-		}
-		// If the HTTP response code is 200 as defined in example definition, your page structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
-		// page = armcomputeschedule.ScheduledActionExtensionClientListByVMsResponse{
-		// 	ScheduledActionResourcesListResult: armcomputeschedule.ScheduledActionResourcesListResult{
-		// 		Value: []*armcomputeschedule.ScheduledActionResources{
-		// 			{
-		// 				Properties: &armcomputeschedule.ScheduledActionsExtensionProperties{
-		// 					ResourceType: to.Ptr(armcomputeschedule.ResourceTypeVirtualMachine),
-		// 					ActionType: to.Ptr(armcomputeschedule.ScheduledActionTypeStart),
-		// 					StartTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.281Z"); return t}()),
-		// 					EndTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.286Z"); return t}()),
-		// 					Schedule: &armcomputeschedule.ScheduledActionsSchedule{
-		// 						ScheduledTime: to.Ptr(func() time.Time { t, _ := time.Parse(time.TimeOnly, "19:00:00"); return t}()),
-		// 						TimeZone: to.Ptr("g"),
-		// 						RequestedWeekDays: []*armcomputeschedule.WeekDay{
-		// 							to.Ptr(armcomputeschedule.WeekDayMonday),
-		// 						},
-		// 						RequestedMonths: []*armcomputeschedule.Month{
-		// 							to.Ptr(armcomputeschedule.MonthJanuary),
-		// 						},
-		// 						RequestedDaysOfTheMonth: []*int32{
-		// 							to.Ptr[int32](15),
-		// 						},
-		// 						ExecutionParameters: &armcomputeschedule.ExecutionParameters{
-		// 							OptimizationPreference: to.Ptr(armcomputeschedule.OptimizationPreferenceCost),
-		// 							RetryPolicy: &armcomputeschedule.RetryPolicy{
-		// 								RetryCount: to.Ptr[int32](17),
-		// 								RetryWindowInMinutes: to.Ptr[int32](29),
-		// 							},
-		// 						},
-		// 						DeadlineType: to.Ptr(armcomputeschedule.DeadlineTypeUnknown),
-		// 					},
-		// 					NotificationSettings: []*armcomputeschedule.NotificationProperties{
-		// 						{
-		// 							Destination: to.Ptr("wbhryycyolvnypjxzlawwvb"),
-		// 							Type: to.Ptr(armcomputeschedule.NotificationTypeEmail),
-		// 							Language: to.Ptr(armcomputeschedule.LanguageEnUs),
-		// 							Disabled: to.Ptr(true),
-		// 						},
-		// 					},
-		// 					Disabled: to.Ptr(true),
-		// 					ProvisioningState: to.Ptr(armcomputeschedule.ProvisioningStateSucceeded),
-		// 				},
-		// 				ID: to.Ptr("/subscriptions/83C27AB3-A7B9-498B-B165-D9440661474F/resourceGroups/myRg/providers/Microsoft.ComputeSchedule/scheduledActions/myScheduledAction"),
-		// 				Name: to.Ptr("lktvylsmyrhgllrxisfdeqd"),
-		// 				Type: to.Ptr("afmfvpvaoczrktku"),
-		// 				SystemData: &armcomputeschedule.SystemData{
-		// 					CreatedBy: to.Ptr("cvryvreuvvjtiamcwhisrt"),
-		// 					CreatedByType: to.Ptr(armcomputeschedule.CreatedByTypeUser),
-		// 					CreatedAt: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.288Z"); return t}()),
-		// 					LastModifiedBy: to.Ptr("supbnksztdbgulxgvfmqvriqdlpirh"),
-		// 					LastModifiedByType: to.Ptr(armcomputeschedule.CreatedByTypeUser),
-		// 					LastModifiedAt: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2025-04-17T00:23:55.288Z"); return t}()),
+		// 					LastModifiedAt: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2026-03-12T02:39:41.641Z"); return t}()),
 		// 				},
 		// 			},
 		// 		},